Various bugfixes for Minigames
I did some bugfixes for the minigames, such as changing their schema for optimization on request of Hanz, or passing the parameters as a state instead of in the url for safety.
As stated some days ago, there is a bug that makes Virtual Sear display the cut as “burned” when it hasn’t been burnt. After some investigating, I think this is due to the levels having a gap between min and max, as in, max for Rare is 59 and min for Medium Rare is 60, so any number between 59 and 60 is invalid. I need to find a fix for this.
I ended up fixing this by turning max values from numbers like 59 to 59.99, the app technically uses very precise decimals, so I stopped this from being taken into consideration by rounding the number. This makes sure that there are no decimals, and even if there fucking are, they would need to be very very precisely out of place.